contacts_api/pyproject.toml
[project]
name = "contacts-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Magnteo",email = "bogdanserg48@gmail.com"}
]
readme = "README.md"
packages = [{include = "contacts_api", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
fastapi = ">=0.116.1,<0.117.0"
uvicorn = ">=0.35.0,<0.36.0"
sqlalchemy = ">=2.0.43,<3.0.0"
psycopg = {extras = ["binary"], version = ">=3.2.10,<4.0.0"}
pydantic = {extras = ["email"], version = ">=2.11.9,<3.0.0"}
psycopg2-binary = ">=2.9.10,<3.0.0"
libgravatar = ">=1.0.4,<2.0.0"
python-jose = {extras = ["cryptography"], version = ">=3.5.0,<4.0.0"}
passlib = {extras = ["bcrypt"], version = ">=1.7.4,<2.0.0"}
python-multipart = ">=0.0.20,<0.0.21"
jinja2 = ">=3.1.6,<4.0.0"
slowapi = "^0.1.9"
cloudinary = "^1.44.1"
python-dotenv = "^1.1.1"

[tool.poetry.group.dev.dependencies]
alembic = "^1.16.5"
sphinx = "^8.2.3"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"